+++ Assets/App/Scripts/Entity/Health/IHealth.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public interface IHealth
 {
     public void TakeHealth(int health);
-    public void TakeDamage(int damage, Action onDeath);
+    public void TakeDamage(float damage, Action onDeath);
 }
