+++ Assets/App/Scripts/Entity/Health/PlayerHealth.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour, IHealth
 {
     [Header("Settings")]
-    [SerializeField] int maxHealth;
-    int currentHealth;
+    [SerializeField] float maxHealth;
+    float currentHealth;
 
     //[Header("References")]
 
@@ -29,7 +29,7 @@ public class PlayerHealth : MonoBehaviour, IHealth
         if (currentHealth > maxHealth) currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int damage, Action onDeath)
+    public void TakeDamage(float damage, Action onDeath)
     {
         currentHealth -= damage;
         if (currentHealth <= 0)
